What a View!
Success! His claws came out of the tree with ease! The hopeful squirrel
continued his climb up. A fear of heights was the last thing on this
young squirrel's mind. The higher and higher he went, the more massive
the world appeared to be. After a fair amount of climbing, the young
squirrel took to a branch that hung just over the tree line of the usual
roaming grounds. How infinitesimal it all looked to the young squirrel!
There was so much more to life that the young squirrel never knew!
Colorful birds, a blue ceiling above him, and even more to explore
beyond!