The Top
The Squirrel scurried up farther to the top. He dared not rest on any of
the branches he passed. He had to complete his quest and get back before
anyone noticed. As he was climbing up, he noticed a bird circling the
top of the great tree. Thinking nothing of it, he climbed all the way to
the top. The squirrel was awestruck at the view! He could see his own
house from there! While blissfully unaware of his surroundings, an owl
snuck up on the young squirrel. The owl hooted a mighty hoot and said,
"Your days are numbered! What are you?". The squirrel sat trembling at
the owl, unaware of his intentions.