Fumbling
The young squirrel replied, "Please don't kill me!".  The owl rolled his
eyes, and repeated the phrase, "Your days are numbered! What are you?".
A disgruntled scowl was smeared on the towering beast. "Why are you
doing this to me?" pleaded the small squirrel. The owl sat silent,
waiting for a proper answer to his question. Was this some sort of game
the squirrel pondered? Why would he be so menacing then? The squirrel
could only blankly stare into the direction of the owl, not daring to
make eye contact.