The Answer
Not wanting to anger the owl further, he cried, "A calendar!". The owl's
temper cooled. A smirk replaced the strong opposition on his face.  The
owl opened his mouth once more, "There you are! I was wondering how long
it might take you to prove yourself to be on this tree. If you would at
all!". The squirrel was confused, but he was just happy to not have the
owl glaring at him. "Now then," the owl questioned, "what do you plan to
accomplish on this tree?". "Nothing really. I was just exploring." The
squirrel squeaked. The owl was a tad puzzled himself now, "Really? I
admire your curiosity! I'm sure you weren't expecting me to approach you
here. Feel free to look around. It is a wonderful tree!".