A Small Tour
The curious squirrel took up the owl on his offer. After all, even
though he was probably being missed by his family, he still made the
incredible journey. Might as well look around. The scenery was as you
would expect.  Branches made a convenient path towards the center of the
tree. Within a web of branches sat a giant birds nest where the owl
slept. The owl was not much of a decorator. Twigs and grass jutted out
of the nest like the defenses of a fortress. The inside of his nest was
a little more polished. It looked like care was put into flattening and
softening up the inside of the nest. grass and leaves were layered on
top of each other. It was a quaint little home. The view is what really
put the place together.